+++ src/libs/ValueSchemaError.ts
@@ -14,7 +14,7 @@ export enum CAUSE
 	MIN_LENGTH = "min-length",
 	MAX_LENGTH = "max-length",
 	PATTERN = "pattern",
-	CONVERT_CASE = "convert-case",
+	CASE_CONVERTER = "convert-case",
 
 	CHECKSUM = "checksum",
 }