+++ src/schemas/string.ts
@@ -2,7 +2,7 @@ import * as ifEmptyString from "../appliers/ifEmptyString";
 import * as ifNull from "../appliers/ifNull";
 import * as ifUndefined from "../appliers/ifUndefined";
 import * as only from "../appliers/only";
-import * as convertCase from "../appliers/string/convertCase";
+import * as caseConverter from "../appliers/string/caseConverter";
 import * as maxLength from "../appliers/string/maxLength";
 import * as minLength from "../appliers/string/minLength";
 import * as pattern from "../appliers/string/pattern";
@@ -12,7 +12,7 @@ import * as type from "../appliers/string/type";
 import {BaseSchema} from "../libs/BaseSchema";
 
 export const STRING = {
-	CONVERT_CASE: convertCase.CONVERT_CASE,
+	CASE_CONVERTER: caseConverter.CASE_CONVERTER,
 	PATTERN: pattern.PATTERN,
 };
 
@@ -23,7 +23,7 @@ type OptionsForString =
 	only.Options<string> |
 	type.Options |
 	trims.Options |
-	convertCase.Options |
+	caseConverter.Options |
 	minLength.Options |
 	maxLength.Options |
 	pattern.Options;
@@ -41,7 +41,7 @@ class StringSchema extends BaseSchema<string>
 			ifEmptyString.applyTo,
 			minLength.applyTo,
 			maxLength.applyTo,
-			convertCase.applyTo,
+			caseConverter.applyTo,
 			pattern.applyTo,
 		]);
 	}