+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* `convertCase` option to `string()`
+* `caseConverter` option to `string()`
 
 ### Changed